# docs/openapi/warehouse.openapi.yaml
openapi: 3.1.0
info:
  title: Warehouse HTTP API
  version: 1.0.0
  description: |
    Warehouse 对外 JSON HTTP API 的正式契约。

    WebDAV (`/dav/*`) 和独立 S3 Endpoint 是协议适配层，不在本 OpenAPI
    规范中逐操作重复定义。WebDAV 请参阅用户使用指南，S3 请参阅
    `docs/S3设计方案.md`。
servers:
  - url: https://warehouse.tidukongjian.com
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/public/share/user/upload:
    put:
      tags: [Directed shares]
      operationId: uploadDirectedShareFile
      summary: 向定向分享目录上传文件
      parameters:
        - {name: shareId, in: query, required: true, schema: {type: string, format: uuid}}
        - {name: path, in: query, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
      responses:
        "200": {$ref: "#/components/responses/MessageResponse"}
        "400": {$ref: "#/components/responses/PlainTextError"}
        "403": {$ref: "#/components/responses/PlainTextError"}
        "410": {$ref: "#/components/responses/PlainTextError"}
    post:
      tags: [Directed shares]
      operationId: uploadDirectedShareFilePost
      summary: 使用 POST 向定向分享目录上传文件
      parameters:
        - {name: shareId, in: query, required: true, schema: {type: string, format: uuid}}
        - {name: path, in: query, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
      responses:
        "200": {$ref: "#/components/responses/MessageResponse"}
        "400": {$ref: "#/components/responses/PlainTextError"}
        "403": {$ref: "#/components/responses/PlainTextError"}
...
  /api/v1/public/uploads/sessions:
    post:
      tags: [Upload sessions]
      operationId: createUploadSession
      summary: 创建可恢复分片上传会话
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateUploadSessionRequest"}
      responses:
        "201":
          description: 上传会话已创建
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UploadSession"}
        "400": {$ref: "#/components/responses/PlainTextError"}
        "403": {$ref: "#/components/responses/PlainTextError"}
        "413": {$ref: "#/components/responses/PlainTextError"}
  /api/v1/public/uploads/sessions/{sessionId}:
    parameters:
      - {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      tags: [Upload sessions]
      operationId: getUploadSession
      summary: 获取上传会话和已上传分片
      responses:
        "200":
          description: 上传会话
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UploadSession"}
        "403": {$ref: "#/components/responses/PlainTextError"}
        "404": {$ref: "#/components/responses/PlainTextError"}
    delete:
      tags: [Upload sessions]
      operationId: abortUploadSession
      summary: 取消上传会话
      responses:
        "200":
          description: 会话已取消
          content:
            application/json:
              schema:
                type: object
                required: [aborted]
                properties:
                  aborted: {type: boolean, const: true}
        "403": {$ref: "#/components/responses/PlainTextError"}
        "404": {$ref: "#/components/responses/PlainTextError"}
  /api/v1/public/uploads/sessions/{sessionId}/parts/{partNumber}:
    put:
      tags: [Upload sessions]
      operationId: uploadSessionPart
      summary: 上传一个原始字节分片
      parameters:
        - {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: partNumber, in: path, required: true, schema: {type: integer, minimum: 1}}
        - name: X-Warehouse-Checksum-SHA256
          in: header
          required: true
          description: 分片内容的 64 位十六进制 SHA-256；也兼容 x-amz-checksum-sha256
          schema: {type: string, pattern: "^[0-9a-fA-F]{64}$"}
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: {type: string, format: binary}
      responses:
        "200":
          description: 分片上传成功
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UploadPartResponse"}
        "400": {$ref: "#/components/responses/PlainTextError"}
        "403": {$ref: "#/components/responses/PlainTextError"}
        "404": {$ref: "#/components/responses/PlainTextError"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or UCAN
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
...
  responses:
...
    PlainTextError:
      description: 当前 Handler 使用 http.Error 返回的纯文本错误
      content:
        text/plain:
          schema: {type: string}
...
    MessageResponse:
      description: 操作成功
      content:
        application/json:
          schema:
            type: object
            required: [message]
            properties:
              message: {type: string}
  schemas:
...
    CreateUploadSessionRequest:
      type: object
      required: [path, size, chunkSize, fileName]
      properties:
        path: {type: string, minLength: 1}
        shareId: {type: string, format: uuid, description: 定向分享上传时填写}
        resourceId: {type: string, description: 收到的 V3 共享资源上传时填写；不可与 shareId 混用}
        size: {type: integer, format: int64, minimum: 0}
        chunkSize: {type: integer, format: int64, minimum: 1}
        fileName: {type: string, minLength: 1}
        contentType: {type: string}
        lastModified: {type: integer, format: int64}
    UploadSessionPart:
      type: object
      required: [partNumber, size, etag, checksumSha256, updatedAt]
      properties:
        partNumber: {type: integer, minimum: 1}
        size: {type: integer, format: int64, minimum: 0}
        etag: {type: string}
        checksumSha256: {type: string}
        updatedAt: {type: string, format: date-time}
    UploadSession:
      type: object
      required: [id, scope, path, size, chunkSize, fileName, status, createdAt, updatedAt, expiresAt, parts]
      properties:
        id: {type: string, format: uuid}
        scope: {type: string}
        path: {type: string}
        shareId: {type: string, format: uuid}
        resourceId: {type: string}
        size: {type: integer, format: int64, minimum: 0}
        chunkSize: {type: integer, format: int64, minimum: 1}
        fileName: {type: string}
        contentType: {type: string}
        lastModified: {type: integer, format: int64}
        status: {type: string}
        createdAt: {type: string}
        updatedAt: {type: string}
        expiresAt: {type: string}
        parts:
          type: array
          items: {$ref: "#/components/schemas/UploadSessionPart"}
    UploadPartResponse:
      type: object
      required: [session, part]
      properties:
        session: {$ref: "#/components/schemas/UploadSession"}
        part: {$ref: "#/components/schemas/UploadSessionPart"}
security:
  - bearerAuth: []